removing utils folder and moving functions to the folders where they belong as utility files
